refactor: combine remote context with datatable context
