+++ lib/src/components/data-table/DataTable.mdx
@@ -1,6 +1,6 @@
 ---
 title: DataTable
-component: DataTable, DataTable.Remote, DataTable.Head, DataTable.HeaderCell, DataTable.Body, DataTable.Row, DataTable.DataCell, DataTable.GlobalFilter, DataTable.Remote, DataTable.Loading, DataTable.Error
+component: DataTable, DataTable.Head, DataTable.HeaderCell, DataTable.Body, DataTable.Row, DataTable.DataCell, DataTable.GlobalFilter, DataTable.Loading, DataTable.Error
 description: Displays tabular data with features such as sorting and pagination
 category: Content
 ---
@@ -11,8 +11,7 @@ category: Content
 
 ## Anatomy
 
-`DataTable` has two different roots namely; `DataTable` and `DataTable.Remote`. Both components manages the table's state and exposes it via the React Context API. This state can be accessed by any child components by calling `useDataTable`. You can pass an `initialState` prop to `DataTable` as described in the [`@tanstack/react-table` docs](https://tanstack.com/table/v8/docs/api/core/table#initialstate).
-`DataTable.Remote` is used when you want achieve server-side pagination and sorting.
+The root DataTable component manages the table's state and exposes it via the React Context API. This state can be accessed by any child components by calling useDataTable. You can pass an initialState prop to DataTable as described in the [`@tanstack/react-table` docs](https://tanstack.com/table/v8/docs/api/core/table#initialstate).
 
 Other `DataTable` components call `useDataTable` and provide useful default implementations for common patterns. For example, `DataTable.Head` will render a header for every column defined in the parent `DataTable`. `DataTable.Body` will render a row for every data item. `DataTable.Table` combines both `DataTable.Head` and `DataTable.Body`.
 
@@ -136,7 +135,7 @@ Note also that `useDataTable` can only be called by a child component of `DataTa
 
 ### Server-side pagination and sorting
 
-`DataTable.Remote` accepts a `fetcher` function that is used to fetch the paginated data. The `fetcher` function provides you with the necessary parameters to query your api endpoint; `pageIndex`, `pageSize`, `sortBy` and `sortDirection`. The response from the `fetcher` function must match the following schema:
+`DataTable` accepts a `fetcher` function that is used to fetch the paginated data. The `fetcher` function provides you with the necessary parameters to query your api endpoint; `pageIndex`, `pageSize`, `sortBy` and `sortDirection`. The response from the `fetcher` function must match the following schema:
 
 ```ts
 {
@@ -145,14 +144,14 @@ Note also that `useDataTable` can only be called by a child component of `DataTa
 }
 ```
 
-A loading state using `<DataTable.Loading>` is automatically included in `DataTable.Remote` which is visible while the `fetcher` promise is pending. You can use `DataTable.Error` to display your own error component when the `fetcher` function promise rejects. It provides a `retry` function, allowing the children of the component to recall `fetcher` with the current pagination and sorting parameters.
+A loading state using `<DataTable.Loading>` is automatically included in `DataTable` which is visible while the `fetcher` promise is pending. You can use `DataTable.Error` to display your own error component when the `fetcher` function promise rejects. It provides a `retry` function, allowing the children of the component to recall the `fetcher` function with either the last pagination options or your own custom pagination properties
 
 ```tsx
-<DataTable.Remote
+<DataTable
   columns={columns}
   defaultPageSize={10}
   defaultSort={{ column: 'name', direction: 'asc' }}
-  fetcher={async (pageIndex, pageSize, sortBy, sortDirection) => {
+  fetcher={async ({ pageIndex, pageSize, sortBy, sortDirection }) => {
     const nextPage = pageIndex + 1
     const data = await fetch(
       `https://your-api?page=${nextPage}&pageSize=${pageSize}&order=${sortBy}&dir=${sortDirection}`
@@ -171,13 +170,15 @@ A loading state using `<DataTable.Loading>` is automatically included in `DataTa
       return (
         <>
           <InlineMessage>Error occured</InlineMessage>
-          <Button onClick={retry}>Retry</Button>
+          <Button onClick={() => retry?.({ pageIndex: 3, pageSize: 10 })}>
+            Retry
+          </Button>
         </>
       )
     }}
   </DataTable.Error>
   <DataTable.Pagination />
-</DataTable.Remote>
+</DataTable>
 ```
 
 ## Features
