+++ lib/src/components/data-table/DataTable.tsx
@@ -11,7 +11,7 @@ import { DataTableLoading } from './DataTableLoading'
 import { DataTableRow } from './DataTableRow'
 import { DataTableTable } from './DataTableTable'
 import { Pagination } from './pagination'
-import { RemoteDataTableProvider } from './RemoteDataTableContext'
+
 type TDataTable = React.FC<React.ComponentProps<typeof DataTableProvider>> & {
   /** Default table body implementation for `DataTable`.
    *
@@ -58,13 +58,6 @@ type TDataTable = React.FC<React.ComponentProps<typeof DataTableProvider>> & {
    *
    * Renders all visible cells as `Table.Cell`. If you need more customisation options,
    * you can build your own implementation with the UI-only `Table` components.
-   */
-
-  Remote: typeof RemoteDataTableProvider
-  /** Alternate context provider for remote data
-   *
-   * Replaces `data` prop with `fetcher`, which must be implemented by the consumer and must return table
-   * data given the current `pageSize`, `pageIndex`, `sortBy` and `sortDirection` values
    */
 
   Row: typeof DataTableRow
@@ -116,7 +109,6 @@ DataTable.DataCell = DataTableDataCell
 DataTable.Head = DataTableHead
 DataTable.HeaderCell = DataTableHeaderCell
 DataTable.Pagination = Pagination
-DataTable.Remote = RemoteDataTableProvider
 DataTable.Row = DataTableRow
 DataTable.GlobalFilter = DataTableGlobalFilter
 DataTable.Table = DataTableTable
